Change type of fileSize in Document to int

+++ signer-cloud-server/src/main/java/com/wultra/signercloud/server/document/Document.java
@@ -53,7 +53,7 @@ public class Document {
 
     private String fileName;
 
-    private long fileSize;
+    private int fileSize;
 
     private long documentContentId;
